+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Fixed
 * Fixed the problem that the URL of Webhook API is different (#202)
 * Fixed some attributes are not updated in advanced search results (#230)
+* Fixed that the entry being created cannot be deleted (#242)
 * Fixed update history of TOP page (#258)
 
 ### Refactored