+++ CHANGELOG.md
@@ -3,12 +3,28 @@
 ## In development
 
 ### Added
+
+### Changed
+
+### Fixed
+
+## v3.3.1
+
+### Fixed
+* Fixed bug AttributeValue.parent_attr is different with child's one (#272)
+
+## v3.3.0
+
+### Added
+* Added the django-replicated library (#166)
+* Added job function that cannot be canceled (#199)
 * Added param of editting user for ldap (#256)
 
 ### Changed
+* Changed to remove complement_attrs when requesting show entry page (#166)
+* Changed the logout from GET to POST (#166)
 * Changed the HTTP method on the entry.export page from GET to POST (#166)
 * Changed not to create tokens with GET user.access_token API (#166)
-* Added job function that cannot be canceled (#199)
 * Changed the behavior of token refresh (#208)
 * Upgrade Django version from v3.2.4 to v3.2.5 (#254)
 * Changed cookie of session to secure attribute, and to return HSTS header (#257)